+++ app/src/main/java/org/othr/flashyplayground/activities/FlashcardTopicActivity.kt
@@ -1,10 +1,13 @@
 package org.othr.flashyplayground.activities
 
+import android.content.DialogInterface
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import com.google.android.material.snackbar.Snackbar
 import org.othr.flashyplayground.R
 import org.othr.flashyplayground.databinding.ActivityFlashcardTopicAddBinding
@@ -20,6 +23,9 @@ class FlashcardTopicActivity : AppCompatActivity() {
 
     // Declaration of MainApp
     lateinit var app :  MainApp
+    // Alert Dialog Builder to confirm deletion of topic
+    lateinit var builder: AlertDialog.Builder
+    var edit: Boolean = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
 
@@ -30,7 +36,19 @@ class FlashcardTopicActivity : AppCompatActivity() {
         binding = ActivityFlashcardTopicAddBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        var edit = false
+        // Initalization of Main App
+        app = application as MainApp
+
+        // Setup Dialog if deleting a topic
+        builder = AlertDialog.Builder(this@FlashcardTopicActivity)
+        builder.setTitle(getString(R.string.title_confirm_delete))
+        builder.setMessage(getString(R.string.message_confirm_delete_topic))
+
+        // Toolbar support
+        binding.toolbarTopicAdd.title = title
+        setSupportActionBar(binding.toolbarTopicAdd)
+
+        edit = false
 
         if (intent.hasExtra("edit_topic")) {
             aTopic = intent.extras?.getParcelable("edit_topic")!!
@@ -39,14 +57,8 @@ class FlashcardTopicActivity : AppCompatActivity() {
             // Reuse string variable
             binding.btnTopicAdd.setText(R.string.btn_changeFlashcard)
             edit = true
-
-            // Toolbar support with deleting feature
-            binding.toolbarTopicAdd.title = aTopic.title
-            setSupportActionBar(binding.toolbarTopicAdd)
         }
 
-        // Initalization of Main App
-        app = application as MainApp
 
         binding.btnTopicAdd.setOnClickListener {
             aTopic.title = binding.topicTitle.text.toString()
@@ -63,29 +75,60 @@ class FlashcardTopicActivity : AppCompatActivity() {
                 else {
                     app.flashcards.updateTopic(aTopic.copy())
                 }
+
+                // Launch back list activity after adding / updating the topic
+                var laucherIntent = Intent(applicationContext, SplashScreenActivity::class.java)
+                startActivity(laucherIntent)
             }
 
-            // Always launch back list activity after adding / updating the topic
-
-            var laucherIntent = Intent(applicationContext, SplashScreenActivity::class.java)
-            startActivity(laucherIntent)
+            else {
+                    Snackbar
+                        .make(it, R.string.message_discardedAddTopic, Snackbar.LENGTH_LONG)
+                        .show()
+            }
 
         }
     }
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.menu_topic_delete, menu)
+        menuInflater.inflate(R.menu.menu_topic_add, menu)
+        // enable / disable delete item with respect to edit mode
+        if (edit) {
+            menu?.findItem(R.id.topic_delete_item)?.setEnabled(true)
+        }
+        else {
+            menu?.findItem(R.id.topic_delete_item)?.setEnabled(false)
+        }
         return super.onCreateOptionsMenu(menu)
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when (item.itemId) {
-            R.id.menu_topic_delete -> {
-                app.flashcards.deleteTopic(aTopic.copy())
-                val launcherIntent = Intent(this, SplashScreenActivity::class.java)
-                startActivity(launcherIntent)
+            R.id.topic_delete_item -> {
+                // yes / no button
+                builder.setPositiveButton("Yes", DialogInterface.OnClickListener { dialog, id ->
+                    app.flashcards.deleteTopic(aTopic.copy())
+                    dialog.cancel() // back to main screen (onResume)
+                    val launcherIntent = Intent(this, SplashScreenActivity::class.java)
+                    startActivity(launcherIntent)
+                })
+
+                builder.setNegativeButton("No", DialogInterface.OnClickListener { dialog, id ->
+                    dialog.cancel()
+                })
+
+                // make dialog happen
+                var alert = builder.create()
+                alert.show()
+
+            }
+
+            R.id.topic_cancel_item -> {
+                Toast.makeText(this, R.string.message_canceled, Toast.LENGTH_LONG).show()
+                finish()
             }
         }
         return super.onOptionsItemSelected(item)
     }
+
 }
